refactor(perl-critic): split native critic contracts (#0000)
