+++ crates/perl-lsp-rs-core/src/tooling/perl_critic/native.rs
@@ -5,435 +5,26 @@
 //! subprocess-backed Perl::Critic adapter and built-in fallback so callers can
 //! migrate rule-by-rule without changing runtime behavior in one large step.
 
-use super::{CriticConfig, Severity, Violation, insertion_range};
+#[cfg(test)]
+use super::CriticConfig;
+use super::{Severity, insertion_range};
 use crate::providers::diagnostics::unreachable_code::check_unreachable_code;
 use perl_parser_core::Node;
 use perl_parser_core::NodeKind;
 use perl_parser_core::position::{Position, Range};
 use perl_pragma::PragmaTracker;
 use perl_semantic_analyzer::scope_analyzer::{IssueKind, ScopeAnalyzer, ScopeIssue};
-use serde::{Deserialize, Serialize};
 
-/// Native critic rule category.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum CriticCategory {
-    /// Syntax-level policy that can be checked from tokens or AST shape.
-    Syntax,
-    /// Semantic policy that needs bindings, scopes, or inferred facts.
-    Semantic,
-    /// Workspace-aware policy that needs cross-file/module facts.
-    Workspace,
-    /// Style convention policy.
-    Style,
-    /// Maintainability or complexity policy.
-    Maintainability,
-    /// Security or unsafe-code policy.
-    Security,
-    /// Documentation or POD policy.
-    Documentation,
-}
+mod native_contract;
+mod native_registry;
+mod native_suppressions;
 
-/// Safety level for an optional native critic fix.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum FixSafety {
-    /// Safe to apply automatically.
-    Safe,
-    /// Useful suggestion, but should require user confirmation.
-    Suggested,
-    /// Diagnostic-only guidance with no automatic edit.
-    ManualOnly,
-}
-
-/// Text edit attached to a native critic fix.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CriticTextEdit {
-    /// Range to replace.
-    pub range: Range,
-    /// Replacement text.
-    pub new_text: String,
-}
-
-/// Optional fix attached to a native critic finding.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CriticFix {
-    /// Human-readable action title.
-    pub title: String,
-    /// How safe the edit is.
-    pub safety: FixSafety,
-    /// Edits needed to apply the fix.
-    pub edits: Vec<CriticTextEdit>,
-}
-
-/// Related span for richer native critic diagnostics.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CriticRelatedInformation {
-    /// Related source range.
-    pub range: Range,
-    /// Explanation for the related span.
-    pub message: String,
-}
-
-/// Native critic finding emitted by a rule.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CriticFinding {
-    /// Stable native rule ID, such as `native.variables.unused_lexical`.
-    pub rule_id: String,
-    /// Rule category.
-    pub category: CriticCategory,
-    /// Finding severity.
-    pub severity: Severity,
-    /// Precise source span for the finding.
-    pub range: Range,
-    /// Short diagnostic message.
-    pub message: String,
-    /// Longer explanation shown in editor/CI details.
-    pub explanation: String,
-    /// Suppression key accepted by inline or file-level suppression handling.
-    pub suppression_key: String,
-    /// Related spans, when useful.
-    pub related: Vec<CriticRelatedInformation>,
-    /// Optional fix.
-    pub fix: Option<CriticFix>,
-}
-
-impl CriticFinding {
-    /// Convert this native finding into the existing violation shape.
-    ///
-    /// This is the bridge that lets native rules flow through current LSP,
-    /// execute-command, and diagnostic consumers while those consumers still
-    /// expect `Violation` values.
-    #[must_use]
-    pub fn to_violation(&self, file: impl Into<String>) -> Violation {
-        Violation {
-            policy: self.rule_id.clone(),
-            description: self.message.clone(),
-            explanation: self.explanation.clone(),
-            severity: self.severity,
-            range: self.range,
-            file: file.into(),
-        }
-    }
-}
-
-/// Scope covered by a native critic suppression directive.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum CriticSuppressionScope {
-    /// Suppression applies to the whole file.
-    File,
-}
-
-/// Parsed native critic suppression directive.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CriticSuppression {
-    /// Suppressed rule ID.
-    pub rule_id: String,
-    /// Scope covered by this directive.
-    pub scope: CriticSuppressionScope,
-    /// Zero-based line where the directive appears.
-    pub line: usize,
-    /// Optional human reason after `--`.
-    pub reason: Option<String>,
-}
-
-/// Parsed native critic suppressions for a source file.
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CriticSuppressionMap {
-    suppressions: Vec<CriticSuppression>,
-}
-
-impl CriticSuppressionMap {
-    /// Parse native critic suppression directives from source text.
-    #[must_use]
-    pub fn from_source(source: &str) -> Self {
-        let suppressions = source
-            .lines()
-            .enumerate()
-            .flat_map(|(line, text)| parse_suppression_line(line, text))
-            .collect();
-
-        Self { suppressions }
-    }
-
-    /// Parsed suppression records.
-    #[must_use]
-    pub fn suppressions(&self) -> &[CriticSuppression] {
-        &self.suppressions
-    }
-
-    /// Whether this map suppresses a native critic finding.
-    #[must_use]
-    pub fn suppresses(&self, finding: &CriticFinding) -> bool {
-        self.suppressions.iter().any(|suppression| {
-            suppression.rule_id == finding.rule_id || suppression.rule_id == finding.suppression_key
-        })
-    }
-}
-
-/// Native critic rule context.
-pub struct CriticContext<'a> {
-    /// Source text being analyzed.
-    pub source: &'a str,
-    /// Parsed AST for the source.
-    pub ast: &'a Node,
-    /// Critic configuration.
-    pub config: &'a CriticConfig,
-}
-
-impl<'a> CriticContext<'a> {
-    /// Build a native critic context.
-    pub fn new(source: &'a str, ast: &'a Node, config: &'a CriticConfig) -> Self {
-        Self { source, ast, config }
-    }
-}
-
-/// Native critic rule interface.
-pub trait CriticRule: Send + Sync {
-    /// Stable native rule ID.
-    fn id(&self) -> &'static str;
-
-    /// Rule category.
-    fn category(&self) -> CriticCategory;
-
-    /// Default severity before profile/config remapping.
-    fn default_severity(&self) -> Severity;
-
-    /// Check the source and append findings.
-    fn check(&self, ctx: &CriticContext<'_>, out: &mut Vec<CriticFinding>);
-}
-
-/// Native critic rule bundle used by receipt and readiness tooling.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum NativeCriticProfile {
-    /// Lower-noise candidate for eventual default/native recommended use.
-    Recommended,
-    /// Every registered native rule, useful for strict audits and rule coverage.
-    Strict,
-}
-
-impl NativeCriticProfile {
-    /// Parse a native critic profile token.
-    #[must_use]
-    pub fn parse(raw: &str) -> Option<Self> {
-        match raw {
-            "recommended" => Some(Self::Recommended),
-            "strict" => Some(Self::Strict),
-            _ => None,
-        }
-    }
-
-    /// Stable profile label for receipts.
-    #[must_use]
-    pub const fn as_str(self) -> &'static str {
-        match self {
-            Self::Recommended => "recommended",
-            Self::Strict => "strict",
-        }
-    }
-}
-
-/// Registry for Rust-native critic rules.
-///
-/// The registry is intentionally small orchestration: it owns rule instances,
-/// runs them against a shared context, and returns their findings in registry
-/// order. Runtime diagnostic wiring can build on this without each caller
-/// needing to know how native rules are stored or executed.
-#[derive(Default)]
-pub struct NativeCriticRegistry {
-    rules: Vec<Box<dyn CriticRule>>,
-}
-
-impl NativeCriticRegistry {
-    /// Create an empty native critic registry.
-    #[must_use]
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Create a registry from prebuilt rules.
-    #[must_use]
-    pub fn with_rules(rules: Vec<Box<dyn CriticRule>>) -> Self {
-        Self { rules }
-    }
-
-    /// Create the default recommended native critic registry.
-    ///
-    /// This is the opt-in bundle for callers migrating from ad hoc built-in
-    /// policy execution to the native rule contract. Keep ordering stable so
-    /// diagnostics and receipts are deterministic.
-    #[must_use]
-    pub fn recommended() -> Self {
-        Self::for_profile(NativeCriticProfile::Strict)
-    }
-
-    /// Create a native critic registry for a named profile.
-    ///
-    /// Runtime diagnostics still use `recommended()` for compatibility. The
-    /// explicit profile entry point lets receipts and readiness checks measure
-    /// lower-noise default candidates without changing the opt-in runtime path.
-    #[must_use]
-    pub fn for_profile(profile: NativeCriticProfile) -> Self {
-        match profile {
-            NativeCriticProfile::Recommended => Self::recommended_profile(),
-            NativeCriticProfile::Strict => Self::strict_profile(),
-        }
-    }
-
-    fn recommended_profile() -> Self {
-        Self::with_rules(vec![
-            Box::new(RequireUseStrictRule),
-            Box::new(RequireUseWarningsRule),
-            Box::new(AssignmentInConditionRule),
-            Box::new(PrintfFormatArityRule),
-            Box::new(DeprecatedDefinedRule),
-            Box::new(UndefComparisonRule),
-            Box::new(StaleDollarAtRule),
-            Box::new(UnreachableCodeRule),
-            Box::new(BarewordFilehandleRule),
-            Box::new(TwoArgOpenRule),
-            Box::new(PipeOpenRule),
-            Box::new(UncheckedOpenCloseRule),
-            Box::new(QxReadpipeRule),
-            Box::new(BacktickExecRule),
-            Box::new(StringEvalRule),
-            Box::new(SystemExecRule),
-        ])
-    }
-
-    fn strict_profile() -> Self {
-        Self::with_rules(vec![
-            Box::new(RequireUseStrictRule),
-            Box::new(RequireUseWarningsRule),
-            Box::new(AssignmentInConditionRule),
-            Box::new(PrintfFormatArityRule),
-            Box::new(DeprecatedDefinedRule),
-            Box::new(UndefComparisonRule),
-            Box::new(StaleDollarAtRule),
-            Box::new(UnreachableCodeRule),
-            Box::new(BarewordFilehandleRule),
-            Box::new(TwoArgOpenRule),
-            Box::new(PipeOpenRule),
-            Box::new(UncheckedOpenCloseRule),
-            Box::new(QxReadpipeRule),
-            Box::new(BacktickExecRule),
-            Box::new(StringEvalRule),
-            Box::new(SystemExecRule),
-            Box::new(UnusedLexicalVariableRule),
-            Box::new(UnusedParameterRule),
-            Box::new(DuplicateParameterRule),
-            Box::new(ParameterShadowsGlobalRule),
-            Box::new(DuplicateLexicalDeclarationRule),
-            Box::new(ShadowedLexicalVariableRule),
-            Box::new(CaptureVarWithoutRegexMatchRule),
-            Box::new(UndeclaredVariableRule),
-            Box::new(UninitializedVariableRule),
-            Box::new(UnquotedBarewordRule),
-            Box::new(RequirePodSectionsRule),
-            Box::new(ProhibitLeadingZerosRule),
-        ])
-    }
-
-    /// Add a rule to the registry.
-    pub fn add_rule(&mut self, rule: Box<dyn CriticRule>) {
-        self.rules.push(rule);
-    }
-
-    /// Number of rules in the registry.
-    #[must_use]
-    pub fn len(&self) -> usize {
-        self.rules.len()
-    }
-
-    /// Whether the registry has no rules.
-    #[must_use]
-    pub fn is_empty(&self) -> bool {
-        self.rules.is_empty()
-    }
-
-    /// Stable IDs for registered rules, in execution order.
-    #[must_use]
-    pub fn rule_ids(&self) -> Vec<&'static str> {
-        self.rules.iter().map(|rule| rule.id()).collect()
-    }
-
-    /// Run all registered rules and return collected findings.
-    #[must_use]
-    pub fn check(&self, ctx: &CriticContext<'_>) -> Vec<CriticFinding> {
-        let mut findings = Vec::new();
-
-        for rule in &self.rules {
-            if !rule_enabled(rule.as_ref(), ctx.config) {
-                continue;
-            }
-            rule.check(ctx, &mut findings);
-        }
-
-        let suppressions = CriticSuppressionMap::from_source(ctx.source);
-        findings
-            .into_iter()
-            .filter(|finding| severity_enabled(finding.severity, ctx.config))
-            .filter(|finding| !suppressions.suppresses(finding))
-            .collect()
-    }
-
-    /// Run all registered rules and return current legacy violation values.
-    ///
-    /// This keeps native rule execution single-sourced while callers migrate
-    /// from `Violation` consumers to richer native finding/code-action data.
-    #[must_use]
-    pub fn check_violations(
-        &self,
-        ctx: &CriticContext<'_>,
-        file: impl Into<String>,
-    ) -> Vec<Violation> {
-        let file = file.into();
-        self.check(ctx).into_iter().map(|finding| finding.to_violation(file.clone())).collect()
-    }
-}
-
-fn rule_enabled(rule: &dyn CriticRule, config: &CriticConfig) -> bool {
-    let id = rule.id();
-    let included = config.include.is_empty() || config.include.iter().any(|policy| policy == id);
-    let excluded = config.exclude.iter().any(|policy| policy == id);
-
-    included && !excluded
-}
-
-fn severity_enabled(severity: Severity, config: &CriticConfig) -> bool {
-    severity as u8 >= config.severity
-}
-
-fn parse_suppression_line(line: usize, text: &str) -> Vec<CriticSuppression> {
-    const NO_CRITIC: &str = "## no critic ";
-    const NO_NATIVE_CRITIC: &str = "## no perl-lsp-critic ";
-
-    let trimmed = text.trim_start();
-    let Some(rest) =
-        trimmed.strip_prefix(NO_CRITIC).or_else(|| trimmed.strip_prefix(NO_NATIVE_CRITIC))
-    else {
-        return Vec::new();
-    };
-
-    let (rules, reason) = rest.split_once("--").map_or((rest, None), |(rules, reason)| {
-        let reason = reason.trim();
-        (rules, (!reason.is_empty()).then(|| reason.to_string()))
-    });
-
-    rules
-        .split(|ch: char| ch == ',' || ch.is_whitespace())
-        .filter_map(|rule_id| {
-            let rule_id = rule_id.trim();
-            (!rule_id.is_empty()).then(|| CriticSuppression {
-                rule_id: rule_id.to_string(),
-                scope: CriticSuppressionScope::File,
-                line,
-                reason: reason.clone(),
-            })
-        })
-        .collect()
-}
+pub use native_contract::{
+    CriticCategory, CriticContext, CriticFinding, CriticFix, CriticRelatedInformation, CriticRule,
+    CriticTextEdit, FixSafety,
+};
+pub use native_registry::{NativeCriticProfile, NativeCriticRegistry};
+pub use native_suppressions::{CriticSuppression, CriticSuppressionMap, CriticSuppressionScope};
 
 /// Native rule that requires a file-level `use strict;` pragma.
 ///
